requirements.txt
numpy==1.18.4
tensorflow==2.2.0
pillow==7.1.2